requirements.txt
pandas==1.5.3
google-generativeai
python-dotenv
pdf2image
surprise